backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & NLP
openai>=1.3.0

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Database
supabase==2.0.3
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1